+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Optional params for redeeming coupons
 - Updated spinners
 - Update resource selects to be label-focused
+- Make plugin execution architecture aware
 
 ## [0.3.0] - 2017-08-29